Split SensorFilter core components into seperate files
